Enable `checkpoint_sync` by default (#3777)
* Enable `checkpoint_sync` by default

Provide fast synchronization by default.

* Add newline to separate fields

Make it slighly easier to read.

* Update `checkpoint_sync` documentation

Change the documentation to match the new default value, and explain
that changing the value can be used for debugging.

* Improve documentation

Remove `post-Canopy`, because Zebra will likely change checkpoint
heights in the future. Also explain better what the checkpoints are and
why it helps with debugging.

Co-authored-by: teor <teor@riseup.net>

Co-authored-by: teor <teor@riseup.net>