feat(rpc): Log unrecognized RPC requests (#3860)
* Trace RPC calls by default

Enable RPC call tracing by default. This is useful for development
purposes, and should probably be removed in the future.

* Create a `TracingMiddleware` for the RPC server

A middleware that will print error messages if there are unrecognized
RPC requests.

* Use `TracingMiddleware` in the `Server`

Enable logging of unrecognized RPC requests.

* Warn user about unrecognized RPCs

Instead of reporting it as an error.

Co-authored-by: teor <teor@riseup.net>

* Revert "Trace RPC calls by default"

This reverts commit 6d7f10de2211b6d5ae581618e799069775717a25.

* Clone `Call` instead of creating a `String`

This might improve performance.

Co-authored-by: teor <teor@riseup.net>