ci(disks): truncate disk names and improve disk name search in PRs
Previous behavior:
Some tests were not able to create a disk image as the following error was being raised
in GCP: Invalid value for field 'resource.name'. Must be a match of regex '(?:[a-z](?:[-a-z0-9]{0,61}[a-z0-9])?)'

Expected behavior:
Disk image names in GCP have a 61 characters max length constraint, and disks names
are being created with a length between 71 and 84 chars.

Solution:
This refactors the disk naming and the way we find the latest cached disk image.
Instead of searching the commit SHA of the current "running" commit, the first
search is focused on finding a disk from the actual PR, as matching the actual
commit will rarely be a match unless re-running a job, and even in this scenario
we can search in the PR and the latest disk will be returned from a successful run

Here's a disk name example with the longest name from a PR after this changes are
applied: `zebrad-cache-4962-merge-v25-mainnet-checkpoint-update` (53 chars long).
If a disk is manually triggered using `workflow_dispatch` the branch name must
be <= 18 characters so we truncate the branch name at disk creation and while
searching for an available disk name for the same branch.