fix(docker): do not mix cache images with test images
Docker does not allow to use multiple tags for cache images in their
`cache-from` and `cache-to` options. Which makes some images to lose
previous tags, if the same.

This cause automated mechanims for deletion to fail, as those cache images
were dependant on the main image, and deleting the cache alone was raising
an error if the main image was not deleted first.

A workaround for this is making a separate repository which can hold this
cache images. Which also reduces clutter from the main registries.