fix(ci): handle `tee` trying to write to a closed pipe
Use `trap '' PIPE` as a command in the pipeline might exit early and we want the script to continue executing, as using `grep --max-count=1` might exit after finding the first match, causing a broken pipe error with `tee`.

This is being combined with  `set -o pipefail` to have strict error handling and we want the script to fail if any command in the pipeline fails.