Fix docs links
PR #7886 introduced this link to docs.rs for the docs of the public
APIs: https://docs.rs/releases/search?query=zebra. However, the query in
the link returns crates that are not part of the Zebra's public API,
and doesn't return the two tower-* crates that are part of Zebra.

This commit refers to a manually created list of links pointing to the
docs of the public APIs of all current Zebra crates.